requirements.txt
alembic==1.4.3
apache-airflow==1.10.12
apispec==1.3.3
argcomplete==1.12.1
astroid==2.4.2
attrs==19.3.0
Babel==2.8.0
cached-property==1.5.2
cattrs==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.35
defusedxml==0.6.0
dill==0.3.2
dnspython==2.0.0
docutils==0.16
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
graphviz==0.14.2
gunicorn==20.0.4
idna==2.10
iso8601==0.1.13
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
natsort==7.0.1
numpy==1.19.2
pandas==1.1.3
pendulum==1.4.4
prison==0.1.3
psutil==5.7.2
Pygments==2.7.1
PyJWT==1.7.1
pylint==2.6.0
pyrsistent==0.17.3
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
pytzdata==2020.1
PyYAML==5.3.1
requests==2.24.0
setproctitle==1.1.10
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
tabulate==0.8.7
tenacity==4.12.0
text-unidecode==1.3
thrift==0.13.0
toml==0.10.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.10
Werkzeug==0.16.1
wrapt==1.12.1
WTForms==2.3.3
zope.deprecation==4.4.0
